backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psutil==5.9.5
aiofiles==23.2.1
python-dotenv==1.0.0
bcrypt==4.0.1
websockets==11.0.3
pyotp==2.9.0
asyncssh==2.14.0
motor==3.3.1
beanie==1.23.0
fastapi-socketio==0.0.10
croniter==1.4.1
python-crontab==3.0.0 